roll back (새 개발환경 적용)
- makefile에서 pthread에 관련된 옵션 순서 변경
- 테스트 클라이언트 접속 소켓 주소 변경(localhost로)
응

+++ test/Main.cpp
@@ -12,7 +12,7 @@ int main()
 	S_PT_SESSION_U data2;
 	READ_PT_SESSION_U(buffer, data2);
 
-	printf("%d\n", data2.sessionId);	
+	printf("%d\n", (int)data2.sessionId);	
 
 	return 0;
 }
